Update version property syntax in dotnet pack command
The code change in the `gh-nuget-package.yml` file modifies the
`dotnet pack` command. Specifically, it changes the way the
version property is passed to the command. The old syntax
`/p:Version="1.1.${{ github.run_number }}"` is replaced with the
new syntax `--property:Version="1.1.${{ github.run_number }}"`.

This update aligns the syntax with the correct or preferred format.